+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Changed offline directories internal path structure ([#1606](https://github.com/neptune-ai/neptune-client/pull/1606))
 - Changed internal directories path structure ([#1606](https://github.com/neptune-ai/neptune-client/pull/1606))
 - Changed format of warning messages ([#1635](https://github.com/neptune-ai/neptune-client/pull/1635))
+- Make `trash_objects()` raise `ProjectNotFound` if project does not exist ([#1636](https://github.com/neptune-ai/neptune-client/pull/1636))
 
 
 ## 1.8.6